fix: Remove portfolio header and footer from CMS dashboard
